graphrag/requirements.txt
certifi==2025.8.3
charset_normalizer==3.4.2
faiss_cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
graphrag==0.0.0
hf_xet==1.1.5
huggingface_hub==0.34.3
idna==3.10
jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip==25.2
python_dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit_learn==1.6.1
sentence_transformers==5.0.0
setuptools==58.0.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
python-dotenv>=1.0.1
# --- MCP server (stdio) compatible with Claude's protocol handshake ---
mcp==1.13.0
anyio>=4.4
pydantic>=2.7
# Optional for semantic search over text_units
# faiss-cpu>=1.7.4
